Style changes to discover
In discover: Removed style changes from javascript happening on creation
of the "no movies div" and added the changes to the CSS.

In discover: Added more contrast to the genre buttons.

Bug fix: The ul element housing the genre buttons was toggleable. It was
noticed when I "missed" one of the genre buttons. It has been fixed so
it is no longer toggleable.